Move users global into agent
